Refactor server to modular routes and services
